+++ daemon/p2p_announce.h
@@ -9,11 +9,6 @@
 
 void setup_p2p_announce(struct lightningd_state *dstate);
 
-/* Handlers for incoming messages */
-void handle_channel_announcement(struct routing_state *rstate, const u8 *announce, size_t len);
-void handle_channel_update(struct routing_state *rstate, const u8 *update, size_t len);
-void handle_node_announcement(struct routing_state *rstate, const u8 *node, size_t len);
-
 /* Used to announce the existence of a channel and the endpoints */
 void announce_channel(struct lightningd_state *dstate, struct peer *peer);
